requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi
cffi==1.15.1
charset-normalizer==3.3.2
crispy-bootstrap4==2022.1
Django==3.2.25
django-cors-headers==4.1.0
django-crispy-forms==2.0
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
importlib-metadata==6.7.0
outcome==1.3.0.post0
packaging==24.0
Pillow==9.5.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2024.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.7
webdriver-manager==4.0.1
wincertstore==0.2
wsproto==1.2.0
zipp==3.15.0
